+++ CHANGELOG.md
@@ -332,6 +332,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `\OxidEsales\EshopCommunity\Core\Exception\LanguageException`
 - `\OxidEsales\EshopCommunity\Core\Controller\BaseController::setClassName()`, use setClassKey()
 - `\OxidEsales\EshopCommunity\Core\Controller\BaseController::getClassName()`, use getClassKey()
+- `\OxidEsales\EshopCommunity\Core\ShopControl::_stopMonitor()`, use stopMonitoring()
 
 ### Fixed
 - Fix not working actions and promotions [#0005526](https://bugs.oxid-esales.com/view.php?id=5526)